Moved SkillDictionary & Categorized Skills
SkillDescription is now a way of adding more descriptive terms to Skill, downside is that this could probably just be integrated into Skill so that might need some thought.

SkillDictionary has moved to it's own file now that I know where this sort of thing belongs.

PrintSkills now categorizes skills and sorts them by points.

.skill lists now have CSS for being blocks, however some work is needed on the sizing and placement of the blocks.
