+++ LegendsViewer/LegendsViewer.csproj
@@ -90,6 +90,7 @@
     <Compile Include="Controls\HTML\ArtifactPrinter.cs" />
     <Compile Include="Controls\HTML\Utilities\BrowserUtil.cs" />
     <Compile Include="Controls\HTML\Utilities\HTMLStyleUtil.cs" />
+    <Compile Include="Controls\HTML\Utilities\SkillDictionary.cs" />
     <Compile Include="Controls\Map\dlgPopulation.cs">
       <SubType>Form</SubType>
     </Compile>
